Move `TestChainSpec` to sync leap test module
